build: use Node.js 22.14.0
I updated the software requirements to include Fast Node Manager for
installing and selecting the supported version of Node.js for the
product.

I created the setup.sh and Setup.bat scripts to automate preparing the
repository for development.

I created .node-version and stored the version number of Node.js 22.14.0
in the file.