ci: validate commit messages
I updated the software requirements to include Fast Node Manager for
installing and selecting the supported version of Node.js for the
product.

I created the setup.sh and Setup.bat scripts to automate preparing the
repository for development.

I created .node-version and stored the version number of Node.js 22.14.0
in the file.

I created the package.json manifest file for the Authentication Service
project. I added the initial metadata for the project to the file.

I added NPM dependencies on commitlint and Husky. I configured the rules
for commitlint. I defined the commit-msg Git hook handler that will use
commitlint to validate that the commit message is formatted per the
Conventional Commits format.

I created the commitlint GitHub Actions workflow. This workflow will
run commitlint on all of the commits in a pull request and for the
latest commit in the main branch to ensure that the commit messages are
properly formatted.